backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-multipart
passlib[bcrypt]
python-dotenv
python-telegram-bot
httpx
slowapi
psycopg2-binary
cryptography
